+++ cocos/ui/UIRichText.cpp
@@ -334,7 +334,7 @@ public:
 
     void endElement(void *ctx, const char *name) override;
 
-    void textHandler(void *ctx, const char *s, int len) override;
+    void textHandler(void *ctx, const char *s, size_t len) override;
 
     
     void pushBackFontElement(const Attributes& attribs);
@@ -754,7 +754,7 @@ void MyXMLVisitor::endElement(void *ctx, const char *elementName)
     }
 }
 
-void MyXMLVisitor::textHandler(void *ctx, const char *str, int len)
+void MyXMLVisitor::textHandler(void *ctx, const char *str, size_t len)
 {
     std::string text(str, len);
     auto color = getColor();
